flir-thermal-inspector/requirements.txt
flask
flask-cors
flyr
pillow
numpy
werkzeug
